Added Patch and Put HTTP Methods

+++ springMVC/src/main/java/com/example/springMVC/services/EmployeeService.java
@@ -4,9 +4,15 @@ import com.example.springMVC.dto.EmployeeDTO;
 import com.example.springMVC.entities.EmployeeEntity;
 import com.example.springMVC.repositories.EmployeeRepository;
 import org.modelmapper.ModelMapper;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.util.ReflectionUtils;
 
+import java.lang.reflect.Field;
 import java.util.List;
+import java.util.Map;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -21,26 +27,54 @@ public class EmployeeService {
         this.employeeRepository = employeeRepository;
         this.modelMapper = modelMapper;
     }
-    public EmployeeDTO getEmployeeById(Long id) {
+    public ResponseEntity<EmployeeDTO> getEmployeeById(Long id) {
         EmployeeEntity emp = employeeRepository.findById(id).orElse(null);
-        return modelMapper.map(emp, EmployeeDTO.class);
+        if(emp == null) return ResponseEntity.notFound().build();
+        return ResponseEntity.ok(modelMapper.map(emp, EmployeeDTO.class));
     }
 
-    public List<EmployeeDTO> getAllEmployees() {
+    public ResponseEntity<List<EmployeeDTO>> getAllEmployees() {
         List<EmployeeEntity> employees = employeeRepository.findAll();
-        return employees.stream()
+        return ResponseEntity.ok(employees.stream()
                 .map(employeeEntity -> modelMapper.map(employeeEntity,EmployeeDTO.class))
-                .collect(Collectors.toList());
+                .collect(Collectors.toList()));
     }
 
-    public EmployeeDTO create(EmployeeDTO newEmp) {
+    public ResponseEntity<EmployeeDTO> create(EmployeeDTO newEmp) {
         EmployeeEntity newEmployeeEntity = modelMapper.map(newEmp,EmployeeEntity.class);
         EmployeeEntity saveEmp = employeeRepository.save(newEmployeeEntity);
-        return modelMapper.map(saveEmp,EmployeeDTO.class);
+        if(saveEmp == null) return ResponseEntity.notFound().build();
+        return ResponseEntity.ok(modelMapper.map(saveEmp,EmployeeDTO.class));
     }
 
-    public String deleteEmployee(Long id) {
+    public ResponseEntity<String> deleteEmployee(Long id) {
+        EmployeeEntity doesExist = employeeRepository.findById(id).orElse(null);
+        if(doesExist == null) return ResponseEntity.notFound().build();
         employeeRepository.deleteById(id);
-        return "Successfully Deleted emp with id: "+ id;
+        return ResponseEntity.ok("Successfully Deleted emp with id: "+ id);
+    }
+
+    public ResponseEntity<EmployeeDTO> updateEmployee(Long empId, EmployeeDTO employee) {
+        EmployeeEntity updateEmpEntity = modelMapper.map(employee,EmployeeEntity.class);
+        EmployeeEntity exists = employeeRepository.findById(empId).orElse(null);
+        if(exists == null) return ResponseEntity.notFound().build();
+        EmployeeEntity updateEmp = employeeRepository.save(updateEmpEntity);
+        return ResponseEntity.ok(modelMapper.map(updateEmp,EmployeeDTO.class));
+
+    }
+
+    public ResponseEntity<EmployeeDTO> editEmployee(Long empId, Map<String, Objects> body) {
+        EmployeeEntity exists = employeeRepository.findById(empId).orElse(null);
+        EmployeeEntity employeeEntity = employeeRepository.findById(empId).get();
+
+        if(exists == null) return ResponseEntity.notFound().build();
+        body.forEach((field, value) -> {
+            Field fieldToBeUpdated = ReflectionUtils.findField(EmployeeEntity.class, field);
+            fieldToBeUpdated.setAccessible(true);
+            ReflectionUtils.setField(fieldToBeUpdated, employeeEntity, value);
+        });
+
+        return ResponseEntity.ok(modelMapper.map(employeeRepository.save(employeeEntity),EmployeeDTO.class));
+
     }
 }
